authentication system finished : accessToken accessible through the whole app
